refactor: maintenace related page&components
