+++ packages/frontend/src/pages/maintenance/[maintenanceId].tsx
@@ -1,26 +1,19 @@
 import { Dashboard } from '@components/dashboard/Dashboard';
-import Maintenance from '@components/Maintenance/Maintenance';
-
+import { MaintenanceBoard } from '@components/Maintenance/MaintenanceBoard';
 import { NextPage } from 'next';
 import { useRouter } from 'next/router';
 
 const MaintenancePage: NextPage = () => {
   const router = useRouter();
-
   const maintenanceId = Number(router.query.maintenanceId);
 
-  if (isNaN(maintenanceId)) {
-    return <div>heheh</div>
-  } else {
-    return (
-      <div className="w-full h-full p-4 bg-zinc-800">
-        {/* <Dashboard> */}
-          <Maintenance maintenanceId={maintenanceId} />
-        {/* </Dashboard> */}
-      </div>
-    )
-  }
-
+  return (
+    <Dashboard>
+      <MaintenanceBoard
+        maintenanceId={maintenanceId}
+      />
+    </Dashboard>
+  );
 };
 
 export default MaintenancePage;
